# apps/workspace/generated/http/openapi.bundled.yaml
openapi: 3.1.2
info:
  title: Univer Workspace API
  version: 0.1.0
  summary: Product HTTP interface for the Univer Workspace collaboration example.
  description: |
    This contract exposes product concepts such as users, spaces, Nodes,
    Resources, trash batches, grants, operations, and worktrees. Univer
    Collaboration protocol endpoints are configured separately and are not
    redefined here.
servers:
  - url: /
security:
  - sessionCookie: []
...
paths:
  /api/session:
    get:
      tags:
        - Session
      operationId: getSession
      summary: Get the current login session.
      security:
        - {}
        - sessionCookie: []
      responses:
        '200':
          description: Current authenticated or anonymous session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionView'
...
  /api/spaces/{spaceId}/nodes:
    parameters:
      - name: spaceId
        in: path
        required: true
        schema:
          type: string
    get:
      security:
        - {}
        - sessionCookie: []
      description: Anonymous readers receive viewer access through Link Sharing or Space public read; inaccessible content returns 404.
      tags:
        - Nodes
      operationId: listSpaceRootNodes
      summary: List the root Nodes of a Space.
      parameters:
        - $ref: '#/components/parameters/CursorParameter'
        - $ref: '#/components/parameters/LimitParameter'
      responses:
        '200':
          description: Node page for the Space root.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodePage'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/nodes:
    post:
      tags:
        - Nodes
      operationId: createNode
      summary: Create a pure organizational Node.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNode'
      responses:
        '201':
          description: Created Node.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/nodes/{nodeId}:
    parameters:
      - name: nodeId
        in: path
        required: true
        schema:
          type: string
    get:
      security:
        - {}
        - sessionCookie: []
      description: Anonymous readers receive viewer access through Link Sharing or Space public read; inaccessible content returns 404.
      tags:
        - Nodes
      operationId: getNode
      summary: Get Node metadata without loading children or content.
      responses:
        '200':
          description: Node metadata and navigation context.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
        - Nodes
      operationId: updateNode
      summary: Atomically rename or move a Node.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchNode'
      responses:
        '200':
          description: Updated Node.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/nodes/{nodeId}/children:
    parameters:
      - name: nodeId
        in: path
        required: true
        schema:
          type: string
    get:
      security:
        - {}
        - sessionCookie: []
      description: Anonymous readers receive viewer access through Link Sharing or Space public read; inaccessible content returns 404.
      tags:
        - Nodes
      operationId: listNodeChildren
      summary: List the direct children of any Node.
      parameters:
        - $ref: '#/components/parameters/CursorParameter'
        - $ref: '#/components/parameters/LimitParameter'
      responses:
        '200':
          description: Direct child Node page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodePage'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/blob-upload-sessions:
    post:
      tags:
        - Blobs
      operationId: createBlobUploadSession
      summary: Reserve stable identities and create a Blob Upload Session.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParameter'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBlobUpload'
      responses:
        '200':
          description: The existing Upload Session for this Idempotency Key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlobUploadSessionEnvelope'
        '201':
          description: A new Upload Session was reserved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlobUploadSessionEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
  /api/blob-upload-sessions/{uploadId}:
    parameters:
      - name: uploadId
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Blobs
      operationId: getBlobUploadSession
      summary: Get current Upload Session state.
      responses:
        '200':
          description: Current Upload Session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlobUploadSessionEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Blobs
      operationId: abortBlobUploadSession
      summary: Abort an unpublished Blob Upload Session.
      responses:
        '204':
          description: Upload Session aborted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/blob-upload-sessions/{uploadId}/content:
    parameters:
      - name: uploadId
        in: path
        required: true
        schema:
          type: string
    put:
      tags:
        - Blobs
      operationId: uploadBlobContent
      summary: Stream the exact reserved Blob bytes into storage.
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '204':
          description: Bytes received and verified.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/blob-upload-sessions/{uploadId}/complete:
    parameters:
      - name: uploadId
        in: path
        required: true
        schema:
          type: string
    post:
      tags:
        - Blobs
      operationId: completeBlobUploadSession
      summary: Atomically publish the Node and Blob Resource.
      responses:
        '200':
          description: Existing result for an already completed Upload Session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceCreateResponse'
        '201':
          description: Blob Resource published.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceCreateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/blob-resources/{resourceId}/content:
    parameters:
      - $ref: '#/components/parameters/BlobResourceId'
    get:
      parameters:
        - $ref: '#/components/parameters/Range'
        - $ref: '#/components/parameters/IfNoneMatch'
      security:
        - {}
        - sessionCookie: []
      description: Anonymous readers receive viewer access through Link Sharing or Space public read; inaccessible content returns 404.
      tags:
        - Blobs
      operationId: getBlobContent
      summary: Read Blob bytes for client-selected preview or other use.
      responses:
        '200':
          $ref: '#/components/responses/BlobBytes'
        '206':
          $ref: '#/components/responses/BlobBytes'
        '304':
          description: The current ETag matches If-None-Match.
        '404':
          $ref: '#/components/responses/NotFound'
        '416':
          $ref: '#/components/responses/RangeNotSatisfiable'
    put:
      tags:
        - Blobs
      operationId: replaceBlobContent
      summary: Replace all Blob bytes while preserving Resource identity.
      description: |
        Publishes immediately without Worktree review. Requires editor access, Content-Length,
        one quoted strong If-Match ETag from the downloaded bytes, and a stable Idempotency-Key.
        Preserves the Node, name, original filename, location, ACL and Resource ID.
        A successful replacement generates a fresh ETag even for identical bytes.
        Replaying a completed key returns its recorded result without writing again.
        Inspect GET /api/operations/{id} after an uncertain response. Pending requests cannot
        be started twice; a failed or interrupted replacement requires a new key.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParameter'
        - name: If-Match
          in: header
          required: true
          schema:
            type: string
          description: Exactly one quoted strong ETag; wildcard and weak validators are not accepted.
        - name: Content-Length
          in: header
          required: true
          schema:
            type: integer
            minimum: 0
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Content replaced, or the recorded result of this completed request.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlobReplacementResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          description: Content changed since the supplied ETag. Download and reconcile before retrying.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: workspace_session
      description: Opaque HttpOnly login session cookie.
    discordBotApiKey:
      type: apiKey
      in: header
      name: x-api-key
      description: Shared secret configured on Workspace and the trusted Discord Bot server.
  schemas:
    Error:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
        message:
          type: string
          description: Human-readable diagnostic text.
        field:
          type: string
          description: Input field related to the error, when applicable.
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/Error'
    User:
      type: object
      additionalProperties: false
      required:
        - id
        - username
        - displayName
        - avatarUrl
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type:
            - string
            - 'null'
          format: uri
...
    SessionView:
      oneOf:
        - $ref: '#/components/schemas/AnonymousSession'
        - $ref: '#/components/schemas/AuthenticatedSession'
...
    NodeSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - spaceId
        - parentNodeId
        - name
        - resource
        - hasChildren
        - updatedAt
        - accessRole
        - capabilities
      properties:
        id:
          type: string
        spaceId:
          type: string
        parentNodeId:
          type:
            - string
            - 'null'
        name:
          type: string
        resource:
          oneOf:
            - $ref: '#/components/schemas/ResourceSummary'
            - type: 'null'
        hasChildren:
          type: boolean
        updatedAt:
          type: string
          format: date-time
        accessRole:
          $ref: '#/components/schemas/AccessRole'
        capabilities:
          $ref: '#/components/schemas/NodeCapabilities'
    NodePage:
      type: object
      additionalProperties: false
      required:
        - space
        - parentNode
        - navigationRootNodeId
        - breadcrumbs
        - nodes
        - nextCursor
      properties:
        space:
          $ref: '#/components/schemas/SpaceSummary'
        parentNode:
          oneOf:
            - $ref: '#/components/schemas/NodeSummary'
            - type: 'null'
        navigationRootNodeId:
          type:
            - string
            - 'null'
        breadcrumbs:
          type: array
          items:
            $ref: '#/components/schemas/Breadcrumb'
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/NodeSummary'
        nextCursor:
          type:
            - string
            - 'null'
    OperationView:
      type: object
      additionalProperties: false
      required:
        - id
        - kind
        - state
        - createdAt
        - updatedAt
        - result
        - error
      properties:
        id:
          type: string
        kind:
          $ref: '#/components/schemas/OperationKind'
        state:
          $ref: '#/components/schemas/OperationState'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        result:
          type:
            - object
            - 'null'
          additionalProperties: true
        error:
          oneOf:
            - $ref: '#/components/schemas/OperationError'
            - type: 'null'
...
    BlobUploadSessionEnvelope:
      type: object
      additionalProperties: false
      required:
        - operation
        - upload
        - uploadTarget
      properties:
        operation:
          $ref: '#/components/schemas/OperationView'
        upload:
          $ref: '#/components/schemas/BlobUploadSession'
        uploadTarget:
          oneOf:
            - $ref: '#/components/schemas/BlobUploadTarget'
            - type: 'null'
...
    AnonymousSession:
      type: object
      additionalProperties: false
      required:
        - authenticated
        - githubOAuthEnabled
        - discordOAuthEnabled
      properties:
        authenticated:
          const: false
        githubOAuthEnabled:
          type: boolean
        discordOAuthEnabled:
          type: boolean
    ExternalIdentitySummary:
      type: object
      additionalProperties: false
      required:
        - provider
        - providerUsername
      properties:
        provider:
          type: string
          enum:
            - github
            - discord
        providerUsername:
          type: string
    AuthenticationMethods:
      type: object
      additionalProperties: false
      required:
        - password
        - externalIdentities
      properties:
        password:
          type: boolean
        externalIdentities:
          type: array
          items:
            $ref: '#/components/schemas/ExternalIdentitySummary'
    AuthenticatedSession:
      type: object
      additionalProperties: false
      required:
        - authenticated
        - githubOAuthEnabled
        - discordOAuthEnabled
        - user
        - authenticationMethods
      properties:
        authenticated:
          const: true
        githubOAuthEnabled:
          type: boolean
        discordOAuthEnabled:
          type: boolean
        user:
          $ref: '#/components/schemas/User'
        authenticationMethods:
          $ref: '#/components/schemas/AuthenticationMethods'
...
    SpaceType:
      type: string
      enum:
        - personal
        - team
    SpaceSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - type
        - name
        - publicRead
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/SpaceType'
        name:
          type: string
        publicRead:
          type: boolean
          description: Whether anyone with a link can browse and open this Space read-only.
...
    UnitType:
      type: string
      enum:
        - sheet
        - doc
        - slide
        - board
        - base
    ResourceCapabilities:
      type: object
      additionalProperties: false
      required:
        - openContent
        - editContent
        - downloadContent
      properties:
        openContent:
          type: boolean
        editContent:
          type: boolean
        downloadContent:
          type: boolean
    UniverResourceSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - kind
        - unitId
        - unitType
        - capabilities
      properties:
        id:
          type: string
        kind:
          type: string
          const: univer
        unitId:
          type: string
        unitType:
          $ref: '#/components/schemas/UnitType'
        capabilities:
          $ref: '#/components/schemas/ResourceCapabilities'
...
    BlobResourceSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - kind
        - mediaType
        - byteSize
        - availability
        - capabilities
      properties:
        id:
          type: string
        kind:
          type: string
          const: blob
        mediaType:
          type: string
        byteSize:
          type: integer
          minimum: 0
        availability:
          $ref: '#/components/schemas/BlobAvailability'
        capabilities:
          $ref: '#/components/schemas/ResourceCapabilities'
    ResourceSummary:
      oneOf:
        - $ref: '#/components/schemas/UniverResourceSummary'
        - $ref: '#/components/schemas/BlobResourceSummary'
      discriminator:
        propertyName: kind
        mapping:
          univer: '#/components/schemas/UniverResourceSummary'
          blob: '#/components/schemas/BlobResourceSummary'
    AccessRole:
      type: string
      enum:
        - owner
        - admin
        - editor
        - viewer
    NodeCapabilities:
      type: object
      additionalProperties: false
      required:
        - browseChildren
        - createChildren
        - rename
        - move
        - trash
        - share
      properties:
        browseChildren:
          type: boolean
        createChildren:
          type: boolean
        rename:
          type: boolean
        move:
          type: boolean
        trash:
          type: boolean
        share:
          type: boolean
    Breadcrumb:
      type: object
      additionalProperties: false
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
    CreateNode:
      type: object
      additionalProperties: false
      required:
        - spaceId
        - parentNodeId
        - name
      properties:
        spaceId:
          type: string
        parentNodeId:
          type:
            - string
            - 'null'
        name:
          type: string
          minLength: 1
          maxLength: 255
    NodeResponse:
      type: object
      additionalProperties: false
      required:
        - node
        - space
        - breadcrumbs
        - navigationRootNodeId
      properties:
        node:
          $ref: '#/components/schemas/NodeSummary'
        space:
          $ref: '#/components/schemas/SpaceSummary'
        breadcrumbs:
          type: array
          items:
            $ref: '#/components/schemas/Breadcrumb'
        navigationRootNodeId:
          type:
            - string
            - 'null'
    PatchNode:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        parentNodeId:
          type:
            - string
            - 'null'
    OperationKind:
      type: string
      enum:
        - createResource
        - createBlobResource
        - replaceBlobContent
        - createWorktree
        - addWorktreeUnit
        - createWorktreeUnit
        - mergeWorktree
        - discardWorktree
        - activateWorktreeResource
    OperationState:
      type: string
      enum:
        - pending
        - completed
        - failed
    OperationError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
...
    ResourceCreateResponse:
      type: object
      additionalProperties: false
      required:
        - operation
        - node
      properties:
        operation:
          $ref: '#/components/schemas/OperationView'
        node:
          $ref: '#/components/schemas/NodeSummary'
...
    CreateBlobUpload:
      type: object
      additionalProperties: false
      required:
        - spaceId
        - parentNodeId
        - name
        - originalFilename
        - byteSize
      properties:
        spaceId:
          type: string
        parentNodeId:
          type:
            - string
            - 'null'
        name:
          type: string
          minLength: 1
          maxLength: 255
        originalFilename:
          type: string
          minLength: 1
          maxLength: 1024
        byteSize:
          type: integer
          minimum: 0
        declaredMediaType:
          type: string
          maxLength: 255
    BlobUploadState:
      type: string
      enum:
        - waitingForUpload
        - uploaded
        - verifying
        - completed
        - failed
        - expired
        - aborted
    BlobUploadSession:
      type: object
      additionalProperties: false
      required:
        - id
        - operationId
        - nodeId
        - resourceId
        - state
        - name
        - originalFilename
        - byteSize
        - receivedSize
        - detectedMediaType
        - sha256
        - expiresAt
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        operationId:
          type: string
        nodeId:
          type: string
        resourceId:
          type: string
        state:
          $ref: '#/components/schemas/BlobUploadState'
        name:
          type: string
        originalFilename:
          type: string
        byteSize:
          type: integer
          minimum: 0
        receivedSize:
          type:
            - integer
            - 'null'
          minimum: 0
        detectedMediaType:
          type:
            - string
            - 'null'
        sha256:
          type:
            - string
            - 'null'
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BlobUploadTarget:
      type: object
      additionalProperties: false
      required:
        - method
        - contentUrl
      properties:
        method:
          type: string
          const: PUT
        contentUrl:
          type: string
    BlobReplacementResult:
      type: object
      additionalProperties: false
      required:
        - operationId
        - resourceId
        - etag
      properties:
        operationId:
          type: string
        resourceId:
          type: string
        etag:
          type: string
...
  parameters:
...
    CursorParameter:
      name: cursor
      in: query
      required: false
      description: Opaque cursor returned by the previous page.
      schema:
        type: string
    LimitParameter:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 100
    IdempotencyKeyParameter:
      name: Idempotency-Key
      in: header
      required: true
      description: Random opaque key representing one user intent.
      schema:
        type: string
        minLength: 16
        maxLength: 200
        pattern: ^[A-Za-z0-9_-]+$
    BlobResourceId:
      name: resourceId
      in: path
      required: true
      schema:
        type: string
    Range:
      name: Range
      in: header
      required: false
      schema:
        type: string
      description: A single RFC 9110 bytes range.
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      schema:
        type: string
      description: Return 304 when the current content ETag matches.
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: A valid login session is required.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthenticated:
              value:
                error:
                  code: UNAUTHENTICATED
                  message: Authentication is required.
    Forbidden:
      description: The current User cannot perform this action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: The resource does not exist or is not discoverable.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value:
                error:
                  code: NOT_FOUND
                  message: The resource was not found.
    Conflict:
      description: The request conflicts with current resource state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    PayloadTooLarge:
      description: The declared Blob size exceeds the configured limit.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RangeNotSatisfiable:
      description: The requested byte range cannot be served.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BlobBytes:
      description: Blob bytes with media type, ETag, disposition, and Range headers.
      headers:
        Accept-Ranges:
          schema:
            type: string
        Content-Range:
          schema:
            type: string
        ETag:
          schema:
            type: string
      content:
        '*/*':
          schema:
            type: string
            format: binary